Processing code potentially 1/2 the time. See notes.
Before, the processing code would first process the latest commit of the
master branch of OneLifeData7, and then use part of that data to
determine the latest released git tag of OneLifeData7, and then do all
the processing again for that version of data.

This commit makes it so you can either do just master, just last
released version, or both (default), using the "edge" and "release"
options. If you don't do edge, it gets the version data from a few quick
git commands, which takes a negligible amount of time.

So, if you just want to process the data for edge or latest release when
debugging, you can do that.